- removed ROCold.java - made several variable names more consistent - renamed Model class to more comprehensible class name TestModelGenerator - adapted ObserverPipeline to use new implementation of ROC class - marked ShowROC in ObserverPipeline as @Deprecated to indicate future removal of this method
